# specs/001-schedule-cook-timer/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Cooking Schedule Planner API
  version: 0.1.0
  description: HTTP endpoints for generating coordinated cooking plans and retrieving baseline food metadata.
servers:
  - url: https://localhost:8443
    description: Local development server (HTTPS preferred for notification permissions)

paths:
  /api/v1/foods:
    get:
      summary: List available food profiles and doneness options
      operationId: listFoods
      responses:
        '200':
          description: Successful retrieval of food profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  foods:
                    type: array
                    items:
                      $ref: '#/components/schemas/FoodProfile'
  /api/v1/plans/generate:
    post:
      summary: Generate a coordinated cooking plan for selected foods
      operationId: generatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanGenerationRequest'
      responses:
        '200':
          description: Plan generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid input payload
        '422':
          description: Business validation failed (e.g., unknown food id, invalid target time)
  /api/v1/plans/recalculate:
    post:
      summary: Adjust an existing plan with updated selections or timing
      operationId: recalculatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanAdjustmentRequest'
      responses:
        '200':
          description: Updated plan with recalculated steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid payload
        '404':
          description: Plan not recognized (client should re-generate)
  /api/v1/plans/{planId}:
    get:
      summary: Retrieve a previously generated plan for session recovery
      operationId: getPlan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan data returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '404':
          description: Plan not found (client should use local fallback)

components:
  schemas:
    FoodProfile:
      type: object
      required: [id, name, category, doneness_levels]
      properties:
        id:
          type: string
          description: Unique slug
        name:
          type: string
        category:
          type: string
          enum: [protein, vegetable, starch, other]
        prep_notes:
          type: string
        default_rest_minutes:
          type: integer
          minimum: 0
        doneness_levels:
          type: array
          items:
            $ref: '#/components/schemas/DonenessOption'
    DonenessOption:
      type: object
      required: [level, cook_minutes]
      properties:
        level:
          type: string
        cook_minutes:
          type: integer
          minimum: 1
        prep_buffer_minutes:
          type: integer
          minimum: 0
          default: 0
    PlanGenerationRequest:
      type: object
      required: [mode, items]
      properties:
        mode:
          type: string
          enum: [start_now, finish_by]
        target_finish_time:
          type: string
          format: date-time
          description: Required when mode is finish_by
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PlanItemInput'
        plan_id:
          type: string
          format: uuid
          description: Optional existing plan identifier for recalculation continuity
    PlanItemInput:
      type: object
      required: [food_id, selected_level]
      properties:
        food_id:
          type: string
        selected_level:
          type: string
        servings:
          type: integer
          minimum: 1
          default: 1
        notes:
          type: string
    PlanAdjustmentRequest:
      type: object
      required: [plan_id]
      properties:
        plan_id:
          type: string
          format: uuid
        remove_food_ids:
          type: array
          items:
            type: string
        update_durations_seconds:
          type: object
          additionalProperties:
            type: integer
            description: Duration adjustment in seconds for the corresponding food_id
        add_steps:
          type: array
          items:
            type: object
            required: [food_id, offset_seconds, duration_seconds]
            properties:
              food_id:
                type: string
              offset_seconds:
                type: integer
                minimum: 0
              duration_seconds:
                type: integer
                minimum: 0
              message:
                type: string
    PlanResponse:
      type: object
      required: [plan]
      properties:
        plan:
          $ref: '#/components/schemas/CookingPlan'
    CookingPlan:
      type: object
      required: [plan_id, mode, items, steps, status, total_duration_seconds]
      properties:
        plan_id:
          type: string
          format: uuid
        mode:
          type: string
          enum: [start_now, finish_by]
        start_time:
          type: string
          format: date-time
        target_finish_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [Draft, Generated, Active, Completed, Cancelled]
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlanItem'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/PlanStep'
        total_duration_seconds:
          type: integer
          minimum: 0
    PlanItem:
      type: object
      required: [food_id, selected_level]
      properties:
        food_id:
          type: string
        selected_level:
          type: string
        servings:
          type: integer
          minimum: 1
        notes:
          type: string
    PlanStep:
      type: object
      required: [sequence_number, food_id, action, offset_seconds, duration_seconds, message]
      properties:
        sequence_number:
          type: integer
          minimum: 1
        food_id:
          type: string
        action:
          type: string
          enum: [start, flip, rest, serve]
        offset_seconds:
          type: integer
          minimum: 0
        duration_seconds:
          type: integer
          minimum: 0
        message:
          type: string